+++ src/metadata/ColumnMetadata.ts
@@ -272,6 +272,8 @@ export class ColumnMetadata {
         }
         if (options.args.options.isArray)
             this.isArray = options.args.options.isArray;
+        if (options.args.options.array)
+            this.isArray = options.args.options.array;
         if (options.args.mode) {
             this.isVirtual = options.args.mode === "virtual";
             this.isParentId = options.args.mode === "parentId";