+++ CHANGELOG.md
@@ -8,8 +8,9 @@ feel free to ask us and community.
 ## 0.1.1 (next)
 
 * added support for `pg-native` for postgres (#975). To use it you just need to install `npm i pg-native` and it will be picked up automatically.
-* now Find Options support -1 and 1 for DESC and ASC values. This is better user experience for MongoDB users. 
+* now Find Options support `-1` and `1` for `DESC` and `ASC` values. This is better user experience for MongoDB users. 
 * now inheritances in embeddeds are supported (#966) 
+* `isArray: boolean` in `ColumnOptions` is deprecated. Use `array: boolean` instead.
 
 ## 0.1.0